delete GTServer proxy and implement its methods in GTProxy directly
